repopulate when module enabled

+++ app/Modules/InventoryMovementsStatistics/src/Jobs/RepopulateLast28DaysTableJob.php
@@ -8,6 +8,7 @@ use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Schema;
 
 /**
  *
@@ -26,20 +27,22 @@ class RepopulateLast28DaysTableJob implements ShouldQueue
      */
     public function handle()
     {
-        DB::table('modules_inventory_movements_statistics_last28days_sale_movements')->truncate();
+        if (Schema::hasTable('modules_inventory_movements_statistics_last28days_sale_movements')) {
+            DB::table('modules_inventory_movements_statistics_last28days_sale_movements')->truncate();
 
-        DB::statement('
-            INSERT INTO modules_inventory_movements_statistics_last28days_sale_movements (
-                inventory_movement_id, sold_at, inventory_id, quantity_sold
-            )
-            SELECT
-                inventory_movements.id as inventory_movement_id,
-                inventory_movements.created_at as sold_at,
-                inventory_movements.inventory_id,
-                inventory_movements.quantity_delta * -1 as quantity_sold
-            FROM `inventory_movements`
-            WHERE inventory_movements.type = "sale"
-            AND inventory_movements.created_at >= DATE_SUB(NOW(), INTERVAL 28 DAY)
-        ');
+            DB::statement('
+                INSERT INTO modules_inventory_movements_statistics_last28days_sale_movements (
+                    inventory_movement_id, sold_at, inventory_id, quantity_sold
+                )
+                SELECT
+                    inventory_movements.id as inventory_movement_id,
+                    inventory_movements.created_at as sold_at,
+                    inventory_movements.inventory_id,
+                    inventory_movements.quantity_delta * -1 as quantity_sold
+                FROM `inventory_movements`
+                WHERE inventory_movements.type = "sale"
+                AND inventory_movements.created_at >= DATE_SUB(NOW(), INTERVAL 28 DAY)
+            ');
+        };
     }
 }
